CDAP-12946 simplify action planning
Simplifications to the action pipeline planner and adding more
unit tests. Modified the Dag split by control logic to handle
actions and conditions a little differently in order to remove
a lot of the complicated logic in the pipeline planner. When
the dag is split by control nodes, it will now make sure all
connected non-control nodes will be placed within the same subdag
and can include multiple actions in the same subdag.