CDAP-13129 ensure a sink is only ever in one mapreduce phase
Fixes a bug in the planner that could cause a sink to be present
in multiple phases of a pipeline. The planner will now insert a
connector in front of a sink if it has multiple reduce nodes
as parents, or if it has a reduce node and a source as a parent.