code cleanup around AppMetadataStore
Cleaning up code that touches AppMetadataStore.
This started out as changing methods that take a ProgramId and
a run id as separate arguments and combining them into a single
ProgramRunId argument. Also added several comments so that the
underlying storage format is documented in the class.

Along the way, also fixed style warnings. Removed guava
Predicate and Function in favor of Java8's version of Predicate
and Function. Used lambda functions where possible, removed
explicit casts that are no longer needed, and removed usages
of deprecated Id classes when encountered.