CDAP-13096 store changes to write cluster state for runs
Adding cluster information to run records. Each run record now
has a 'cluster' field that includes the cluster information. In
addition, adding states for provisioning, provisioned,
deprovisioning, and deprovisioned to the AppMetadataStore.

Also adding builders to RunRecord and RunRecordMeta since there
are a bunch of fields, many of which are nullable. Also because
we commonly create a new record from an existing one and modify
just a couple of fields. This will also make it easier to add
profile to the records once we have them.