CDAP-13217 remove app spec check when recording run state
When we are recording run record state, we should not check for
existence of the application or program. Otherwise, there can
be a situation where an app is modified to remove a running
program. If that program is killed on the cluster, then the run
record will be in the 'running' state, which will eventually
be fixed by the run record corrector. However, when it is time
to save the run state, the program will not be found and the
state update will be ignored. This will loop infinitely.