Fix bugs and address review comments
Fixed a bug where the ProgramDescriptor was not getting stored in
program start event, causing problems when a program is started.

Fixed a bug where programs within a workflow were getting started
twice. Once by the workflow driver and once by the program event
subscriber.

Fixed a race condition in a unit test and fix a unit test
that was failing due the profile always being added as a user
argument.

Fixed a bug where IntegrationTestBase would fail because the
default provisioner does not exist, since there is no extensions
directory to load it from. Changed it so that the default provisioner
is always created in the code instead of loaded as an extension.

Removed ProvisionerStore in favor of ProvisioningService directly
controlling transactions.