CDAP-13309 fix bug in run record corrector that would skip records
The run record corrector was assuming that it reads run records
ordered by start time, but that is an invalid assumption. This
can cause it to skip run records in between batch calls.

Fixed it to always scan the entire time range, since that is not
any less performant than restricting the time range.