fix some flaky tests
Fix the ProgramNotificationSubscriberServiceTest to emulate what
happens when a preview run starts in order to avoid a race where
the program is started, then fails, then the state is updated.
This started failing now because the provisioning service is now
hooked up, which will trigger events such that a program run will
try to be started.

Also fix the MetricsClientTest that was relying on a sleep to make
sure metrics are available.