CDAP-13774 add remote hadoop provisioner
This provisioner just connects to an existing cluster instead of
actually creating and deleting a cluster. It needs to be configured
with the host to connect to, as well as the ssh user and key
required to connect.

As part of this work, moved the standard node properties that CDAP
expects into the spi instead of in the dataproc module. Also fixing
a typo to change the dataproc name from gce-dataproc to gcp-dataproc.