CDAP-13965 fix deprovision of a null cluster
Fixed a null pointer exception that would occur when a cluster
cannot be created. When a cluster cannot be created, the cluster
field in the task state is null. When the provisioning service
is asked to deprovision the cluster, it will pass the null cluster
to the provisioner, which results in a null pointer exception
when the provisioner tries to access any methods.

In these cases, task state should still be deleted,
but the provisioner does not actually need to be called.