CDAP-13887 fix metrics skip on provision failure
Fixed a bug where a failure to provision a cluster would not
increment the failed program runs metric for that profile.
The bug is due to the fact that transition of the program run
state from pending to failed was being implicitly handled by
the AppMetadataStore when it processed a message to store
provisioning state.

Fixed the bug by having the provision task explicitly write
a failed program state instead of relying on the store to handle
the state transition. This allows the ProgramNotificationSubscriber
to retain its existing logic of incrementing counters when it
notices that a program run transitions to an end state.