CDAP-14236 fix and cleanup dataproc property validation
Fixed a bug where the provisioner would attempt to create
credentials for the unsubstituted macro for service account key.
Ideally the account key would be validated if it is not a macro,
but there that would require additional features in the provisioner
framework. For now, ensure that the account key is not validated
unless it is empty. This mirrors the behavior in the last release.

Also performed some cleanup to move logic to lookup the project,
network, and zone from the client into the configuration. This
ensures that the project, network, and zone fields in the conf
are always set and always return non-null values.