CDAP-14022 add more bootstrap step types
Add implementations for the rest of the bootstrap step types for
creating applications, starting programs, and setting system
preferences.

Updated the bootstrap files to create the dataprep application
and to start the service in sandbox. Removed the internal
services that were responsible for creating and starting dataprep
in sandbox.

Also added the bootstrap rest api, which simply reloads the
config and runs all the steps.