CDAP-14161 fix cancel provisioning task
Provisioner task cancellation had a bug where the task thread
could save state after it was cancelled, overwriting the
task state from cancelled to something else.

In general, the tasks also would not quit running as early as
possible. Fixed ProvisioningTask so that it will never successfully
persist task state after it has been cancelled, and so that it
will not execute the next subtask after it has been interrupted.