CDAP-14260 endpoint to get the latest run for multiple programs
Added an endpoint to get the latest run for a batch of programs.
The endpoint follows the same format as the other batch endpoints.

Did some code cleanup to move runs lookup for a program from the
handler into the ProgramLifecycleService. This refactoring fixed
a bug where auth checks were not being applied to the program
runs endpoint.