CDAP-14304 improve dataproc properties
Make the dataproc properties more user-friendly.

Removing region as a configurable property in favor of always
using 'global'. This is because region is broken. It will be
re-introduced whenever CDAP-14376 is fixed.

Changed all the optional properties to support an 'auto-detect'
special value that can also be used instead of leaving the property
empty, as we have found that users generally do not expect to
leave fields blank.