CDAP-14562 fixed a classloading issue with output formats
Instead of just using the output format's classloader as the
context classloading when creating the RecordWriter, use a
CombineClassLoader of the output format's classload and the current
context classloader. This prevents issues where classes in the
Hadoop classpath are not visible when creating the RecordWriter.

More specifically, this fixes issues with XML parsers required
by the S3 filesystem. If the program/plugin does not have xml
dependencies packaged (they shouldn't need to), it should not
result in ClassNotFound exceptions.