CDAP-14708 fix a provisioning error for firewall deny rules
Fixed a bug in the dataproc provisioner that would cause
provisioning to fail if the project has a firewall rule that
denies traffic, regardless of whether that rule is relevant or not.
The fix is to check for null instead of assuming an empty list
is returned from the allowed list.