CDAP-14686 pipeline stage validation endpoint
Added an InvalidConfigPropertyException that plugins can throw
to indicate that a specific field has a problem. Also added an
InvalidStageException that can capture multiple errors at once
so that users can fix them all at once instead of one by one.

Refactored the PipelineSpecGenerator so that it can be used at
configure time by the workflow for planning purposes and at
runtime by the Service for validation purposes. The spec generator
now captures exceptions thrown during configuration and
translates them into error objects that are returned through
the REST API.

Also refactored the PluginConfigurer and PluginContext APIs so
that they throw a specific type of exception when the PluginConfig
can't be created from the PluginProperties given by the user.