CDAP-14905 Added main class for system app service programs
Added a UserServiceProgramMain that can be used to run
Service programs from the command line. The implementation assumes
there is a way to generate a file for the app spec and for the
program options before starting the program.

When the program starts, it localizes the application jar and
any required plugin jars in order to ensure the program classloader
has the classes required to instantiate its classes.