CDAP-15047 improve bootstrap steps
Fixed a bug in the program starter where provisioning steps were
being skipped.

Added a log message when the program starter doesn't due anything
because the program is not stopped.

Enhanced the system artifacts loader so that system artifacts
are loaded in parallel. This speeds up bootstrap quite a bit
in environments where the filesystem is slower.