CDAP-15299 fix structured record sorting
Fix a bug in aggregators run in mapreduce pipelines where group
keys would get dropped when different records hash to the same
value.

Implemented a proper comparison of records instead of comparing
their hash codes.

Also enforce that union schemas do not contain multiple maps
or multiple arrays. This conforms with the
Avro schema specification. Without this restriction, it is not
possible to tell which schema in the union a collection matches
to without examining all the data in the collection.