CDAP-15742 add safeguards to prevent cloud runs from getting stuck
Configured the retry strategy to default to retry fetching monitor
messages for one hour. If the monitor times out, it will kill the
remote process and transition the program to the failed state.

On the other side, added a timeout to the amount of time the
RuntimeMonitorServer will wait while shutting down. If it doesn't
get a shutdown message from the master for that same amount of time,
it will shut itself down.