(CDAP-16610) Fixing DataprocUtils.getBucketName null pointer exception
Why:
Starting a pipline failed with the null pointer exception.

What:
This bucket is used for launching jobs in dataproc. Its name is set
via provisioner.system.properties.gcp-dataproc.bucket which is in
system context. But we failed to propagate it to the context
when setting one up for runtime job manager