[CDAP-16975] Default to latest version while adding plugin from left panel +  [CDAP-16976] Reset the default plugin store post CDAP upgrade
This is to ensure we have the following behavior,

1. Pre-upgrade - When the user chooses a plugin from side panel in studio we
choose the latest version and set that as default.

2. Pre-upgrade - When the user consciously changes to a older version of the plugin we
default to that specific version after that.

3. Post-upgrade - Once the user upgrades the CDAP platform, we no longer want them to use
the default plugin version they were using in the previous version of CDAP. So we reset
the default plugins and start afresh

4. Post-upgrade, Post-reset - After reset post upgrade we then follow the pre-upgrade behavior
where from then on we choose the latest version of the plugin for the first time and default
to that for subsequent usage.

As part of that we now store the current version of CDAP in the user store. Post upgrade
there will be a mis-match between the cdap version in the user store and the current cdap
version the user is running, indicating that there was an upgrade upon which we then reset.