CDAP-17078 implement stage consolidation for streaming pipelines
Implemented stage consolidation for streaming pipelines.
Most of the logic is in a new StreamingMultiSinkFunction that
is the multi-sink counterpart to StreamingBatchSinkFunction.

Instead of instantiating a single sink, preparing it, running it,
then finishing it, the multi function prepares all stages
in the group, executes the group using the existing
MultiSinkFunction, then finishes each stage.