CDAP-17202 perform MultiOutputCommitter duties in parallel
setup, abort, recover, and commit jobs and tasks in parallel in
the MultiOutputCommitter.

Also added a tiny performance gain and debuggability improvement
by avoiding a union with an empty RDD in sources. This is mostly
useful for avoiding confusiong when looking at DAGs in the Spark
UI, where jobs would always start with a union between an empty
RDD and the real input RDD.