CDAP-17559 remove Spark adaptive excecution default in app
adaptive execution is meant for Spark3, but it also affects Spark2.
Since it hasn't been extensively tested with Spark2 pipelines,
removing the default in the app and relying on the provisioner
to set depending on the Spark version.

For dataproc, the images with Spark3 already default the setting
to true.