Added logic to return running records as of the start of the specified window for the dashboard.
For performance reasons, we only look back to the last 30 days of activity.

If there are updates in the window for any of the records mentioned above, they will be removed so only the latest update is displayed.

In order to ensure a consistent behavior, records will be sorted based on start timestamp.

Updated Hearbeat table to use MultiScan instead of scanning for each individual namespace